// src/server/graphql/generated/client.ts
export type GetActivitiesQuery = { activities: { nodes: Array<{ id: string, status: string, title: string, slug: string, featuredImageId: string, dateGmt: string, contentTypeName: string, modifiedGmt: string, content: string, featuredImage: { node: { altText: string, id: string, fileSize: number, dateGmt: string, mediaItemUrl: string, mediaType: string, mimeType: string, modifiedGmt: string, slug: string, sourceUrl: string, mediaDetails: { width: number, height: number, file: string } } }, activityDetails: { address: string, place: string, startDatetime: string, geo: { latitude: number, longitude: number, streetAddress: string } }, categories: { nodes: Array<{ name: string, id: string }> } }> } };

export type GetProjectsEnergyQueryVariables = Exact<{ [key: string]: never; }>;


export type GetProjectsEnergyQuery = { projectsEnergy: { nodes: Array<{ content: string, dateGmt: string, featuredImageId: string, id: string, modifiedGmt: string, title: string, slug: string, status: string, projectEnergyDetails: { order: number, geo: { latitude: number, longitude: number, streetAddress: string } }, featuredImage: { node: { altText: string, dateGmt: string, id: string, mediaType: string, mimeType: string, modifiedGmt: string, slug: string, sourceUrl: string, fileSize: number, mediaDetails: { height: number, width: number, file: string } } } }> } };